blockchain_parameters in eosiolib updated to be consistent with native side #2682

+++ contracts/eosiolib/privileged.hpp
@@ -10,18 +10,18 @@ namespace eosio {
       uint32_t target_block_net_usage_pct;
       uint32_t max_transaction_net_usage;
       uint32_t base_per_transaction_net_usage;
-      uint64_t context_free_discount_net_usage_num;
-      uint64_t context_free_discount_net_usage_den;
+      uint32_t context_free_discount_net_usage_num;
+      uint32_t context_free_discount_net_usage_den;
 
-      uint64_t max_block_cpu_usage;
+      uint32_t max_block_cpu_usage;
       uint32_t target_block_cpu_usage_pct;
       uint32_t max_transaction_cpu_usage;
       uint32_t base_per_transaction_cpu_usage;
       uint32_t base_per_action_cpu_usage;
       uint32_t base_setcode_cpu_usage;
       uint32_t per_signature_cpu_usage;
-      uint64_t context_free_discount_cpu_usage_num;
-      uint64_t context_free_discount_cpu_usage_den;
+      uint32_t context_free_discount_cpu_usage_num;
+      uint32_t context_free_discount_cpu_usage_den;
 
       uint32_t max_transaction_lifetime;
       uint32_t deferred_trx_expiration_window;
